[#565] Fix date filters in Firefox
The issue is that we use a polyfill on Firefox and other browsers that have no
`input[type="date"]`. Our polyfill should always send the dates in format
YYYY-MM-DD, but it's not the case. I filled an issue on their repository at
https://github.com/brianblakely/nodep-date-input-polyfill/issues/31. Meanwhile,
we now also accept dates in the format MM/DD/YYYY, which is the EN locale.

Fixes opentrials/opentrials#565